small nit
